Refactor: view (printView) 분리
