+++ src/utils/WinningLotto.js
@@ -1,8 +1,5 @@
 import { Console } from "@woowacourse/mission-utils";
-import { CONSOLE_MESSAGE } from "../Constants.js";
-import Lotto from "../errors/Lotto.js";
-import Bonus from "../errors/Bonus.js";
-import { InputView } from "../view/inputVeiw.js";
+import { InputView } from "../view/InputVeiw.js";
 class WinningLotto {
   async getWinningLotto() {
     let mainLotto = "";
